+++ protostar/protostar_toml/protostar_contracts.py
@@ -2,8 +2,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-from protostar.protostar_toml.sections._protostar_toml_section import (
-    ProtostarTOMLSection,
+from protostar.protostar_toml._protostar_toml_section import ProtostarTOMLSection
+from protostar.protostar_toml.protostar_toml_reader import ProtostarTOMLReader
+from protostar.protostar_toml.protostar_toml_exceptions import (
+    InvalidProtostarTOMLException,
 )
 
 
@@ -14,6 +16,15 @@ class ProtostarContracts(ProtostarTOMLSection):
     @staticmethod
     def get_name() -> str:
         return "contracts"
+
+    @classmethod
+    def from_protostar_toml(
+        cls, protostar_toml: ProtostarTOMLReader
+    ) -> "ProtostarContracts":
+        section_dict = protostar_toml.get_section(cls.get_name())
+        if section_dict is None:
+            raise InvalidProtostarTOMLException(cls.get_name())
+        return cls.from_dict(section_dict)
 
     @classmethod
     def from_dict(cls, raw_dict: Optional[Dict[str, Any]]) -> "ProtostarContracts":
