--- protostar/protostar_toml/sections/__init__.py
@@ -1,6 +1,0 @@
-from protostar.protostar_toml.sections._protostar_toml_section import (
-    ProtostarTOMLSection,
-)
-from protostar.protostar_toml.sections.protostar_config import ProtostarConfig
-from protostar.protostar_toml.sections.protostar_contracts import ProtostarContracts
-from protostar.protostar_toml.sections.protostar_project import ProtostarProject
